ref(nextjs/sveltekit): Move code templates into subdirectories
